requirements.txt
asgiref==3.2.3
Django==3.0.3
django-bootstrap-form==3.4
django-stdimage==5.0.3
Pillow==7.1.1
progressbar2==3.50.1
python-utils==2.4.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
virtualenv==16.7.8
virtualenvwrapper-win==1.2.5
